PFP-6661 Beregningsgrunnlag mapper

+++ dokumentlager/domene/src/main/java/no/nav/foreldrepenger/melding/beregningsgrunnlag/BeregningsgrunnlagPeriode.java
@@ -1,5 +1,6 @@
 package no.nav.foreldrepenger.melding.beregningsgrunnlag;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -8,17 +9,17 @@ import java.util.List;
 import no.nav.foreldrepenger.melding.typer.DatoIntervall;
 
 public class BeregningsgrunnlagPeriode {
-    private int dagsats;
-    private int bruttoPrÅr;
+    private Long dagsats;
+    private BigDecimal bruttoPrÅr;
     private String periodeÅrsak; //Kode beregningsgrunnlagPeriodeÅrsaker.periodeÅrsak
     private DatoIntervall periode;
     private List<BeregningsgrunnlagPrStatusOgAndel> beregningsgrunnlagPrStatusOgAndelList = new ArrayList<>();
 
-    public int getDagsats() {
+    public Long getDagsats() {
         return dagsats;
     }
 
-    public int getBruttoPrÅr() {
+    public BigDecimal getBruttoPrÅr() {
         return bruttoPrÅr;
     }
